Add nav data flash from GPX to device
Currently, only --flash with --erase is implemented.

Writing routes to the device is believed to work, but not well tested.

The HX870 implements the current and previously used navigation targets as indices to the waypoint/route lists. After replacing these lists with new data, those indices are no longer usable and may put the radio in an inconsistent state. This change works around that issue by clearing the navigation status and history for data read from GPX.