Filter duplicate points before flashing
The device implements routes as lists of waypoints. When routes stored in the device are dumped to GPX, it isn't possible to automatically make a distiction between waypoints that the user only added as vertices to define the route, and waypoints that the user also wants to be able to navigate to individually (like any other waypoint). Therefore, all waypoints used in routes must be included in the GPX with both a <wpt> element as well as a <rtept> element.

This change is an attempt to more accurately reproduce the device's memory in the event the GPX file is flashed back to the device: All points (<wpt> or <rtept>) that exactly match another point (in name, position, and comment) are treated as duplicates and filtered from the output. If necessary, routes are modified to use those points that remain after filtering.

This change is not a complete fix. It makes it impossible to exactly reproduce a waypoint list that intentionally includes exact duplicates. However, this is hopefully an exceedingly rare case.